component creation + folgepage rename
